added openrocket logging. realized i put things in the wrong directory
