+++ drivers/sensor/openrocket/imu/CMakeLists.txt
@@ -7,4 +7,4 @@ zephyr_library()
 
 zephyr_library_sources(openrocket_imu.c)
 
-zephyr_include_directories(.)
+# zephyr_include_directories(.)
