+++ drivers/sensor/openrocket/imu/openrocket_imu.c
@@ -1,3 +1,1 @@
 #include "openrocket_imu.h"
-
-int or_data_interpolator() { return sizeof(struct or_data_t); }
