+++ dts/bindings/sensor/openrocket,imu.yaml
@@ -5,7 +5,7 @@ description: IMU that takes its inputs from an openrocket simulation CSV
 
 compatible: "openrocket,imu"
 
-include: [sensor-device.yaml]
+include: [openrocket-sensor-device.yaml]
 
 properties:
   vertical-axis:
@@ -27,15 +27,4 @@ properties:
       - "NEG_X"
       - "NEG_Y"
       - "NEG_Z"
-  sampling-period-us:
-    type: int
-    required: true
-    default: 0
-    description: |
-      What frequency does the sensor update at. The simulated sensor will 
-      only update at this rate no matter how the underlying data changes 
-      beneath it or how often the use calls `sensor_channel_fetch`.
-      A value of 0 (the default) signifies the the value will interpolated 
-      to read as fast as `sensor_channel_fetch` is called.
-# Which axis is the vertical axis
 # What method to use to map 1d lateral acceleration to the remaining 2 axes
