+++ dts/bindings/vendor-prefixes.txt
@@ -1,0 +1,1 @@
+openrocket	Device tree bindings that get their data from openrocket simulation CSVs
